Time: 17 ms (76.56%), Space: 43.8 MB (70.74%) - LeetHub

+++ 0018-4sum/0018-4sum.java
@@ -8,24 +8,17 @@ class Solution {
         Arrays.sort(nums);
         
         for (int i = 0; i < nums.length - 3; i++) {
-            if (i > 0 && nums[i] == nums[i - 1]) continue; // Skip duplicate for `i`
-            
+            if (i > 0 && nums[i] == nums[i - 1]) continue;
             for (int j = i + 1; j < nums.length - 2; j++) {
-                if (j > i + 1 && nums[j] == nums[j - 1]) continue; // Skip duplicate for `j`
-                
+                if (j > i + 1 && nums[j] == nums[j - 1]) continue;
                 int k = j + 1;
                 int l = nums.length - 1;
-                
                 while (k < l) {
-                    int sum = nums[i] + nums[j] + nums[k] + nums[l];
-                    
+                    long sum = (long) nums[i] + nums[j] + nums[k] + nums[l];
                     if (sum == target) {
                         ans.add(Arrays.asList(nums[i], nums[j], nums[k], nums[l]));
-                        
-                        // Skip duplicates for `k` and `l`
                         while (k < l && nums[k] == nums[k + 1]) k++;
                         while (k < l && nums[l] == nums[l - 1]) l--;
-                        
                         k++;
                         l--;
                     } else if (sum < target) {
